requirements.txt
pyyaml==5.1
requests==2.21.0
Jinja2==2.10